backend_real/requirements.txt
aerich==0.8.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
asyncclick==8.1.8.0
asyncpg==0.30.0
bcrypt==4.0.1
click==8.1.8
colorama==0.4.6
dictdiffer==0.9.0
fastapi==0.115.11
h11==0.14.0
httptools==0.6.4
idna==3.10
iso8601==2.1.0
lxml==5.3.1
passlib==1.7.4
pydantic==2.11.0b1
pydantic-settings==2.8.1
pydantic_core==2.31.1
pypika-tortoise==0.5.0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
sniffio==1.3.1
starlette==0.46.1
tomlkit==0.13.2
tortoise-orm==0.24.2
typing-inspection==0.4.0
typing_extensions==4.12.2
uvicorn==0.34.0
watchfiles==1.0.4
websockets==15.0.1
